Restructure, delete sildapiano.apps.accounts directory
