--- sildapiano/sildapiano/apps/accounts/urls.py
@@ -1,20 +1,0 @@
-from django.contrib import admin
-from django.urls import path
-from django.contrib.auth import views as auth_views
-from . import views
-
-app_name = "accounts"
-urlpatterns = [
-    path(
-        "profile/",
-        views.ProfileView.as_view(template_name="accounts/profile.html"),
-        name="profile",
-    ),
-    # Django auth
-    path(
-        "login/",
-        auth_views.LoginView.as_view(template_name="accounts/login.html"),
-        name="login",
-    ),
-    path("logout/", auth_views.LogoutView.as_view(), name="logout"),
-]
