--- sildapiano/sildapiano/apps/accounts/views.py
@@ -1,6 +1,0 @@
-from django.views.generic.base import TemplateView
-from django.contrib.auth.mixins import LoginRequiredMixin
-
-
-class ProfileView(LoginRequiredMixin, TemplateView):
-    template_name = "account/profile.html"
