+++ sildapiano/sildapiano/apps/public/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-
+from django.contrib.auth import views as auth_views
 from . import views
 
 app_name = "public"
@@ -8,5 +8,8 @@ urlpatterns = [
     path("lessons/", views.lessons_page, name="lessons"),
     path("lessons/contacts/", views.contacts, name="contacts"),
     path("lessons/about/", views.about, name="about"),
-    path("lessons/calendar", views.calendar, name="calendar")
+    path("lessons/calendar", views.calendar, name="calendar"),
+    path("profile/", views.ProfileView.as_view(template_name="accounts/profile.html"), name="profile"),
+    path("login/", auth_views.LoginView.as_view(template_name="accounts/login.html"), name="login"),
+    path("logout/", auth_views.LogoutView.as_view(), name="logout"),
 ]
