+++ sildapiano/sildapiano/settings.py
@@ -127,6 +127,6 @@ STATICFILES_DIRS = (os.path.join(PROJECT_DIR, "static"),)
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 
-LOGIN_URL = "accounts:login"
-LOGIN_REDIRECT_URL = "accounts:profile"
+LOGIN_URL = "public:login"
+LOGIN_REDIRECT_URL = "public:profile"
 LOGOUT_REDIRECT_URL = "public:lessons"
