+++ sildapiano/sildapiano/templates/navbar.html
@@ -10,25 +10,25 @@
     <div class="middle">
         <ul>
             {% url 'public:lessons' as lessons_url%}
-            <li>{% if request.path == lessons_url %}Sākums{% else %}<a href={{ lessons_url }}>Sākums</a>{% endif %}</li>
+            <li>{% if request.path == lessons_url %}Sākums{% else %}<a href="{{ lessons_url }}">Sākums</a>{% endif %}</li>
             {% url 'public:about' as about_url%}
-            <li>{% if request.path == about_url %}Par skolotāju{% else %}<a href={{ about_url }}>Par skolotāju</a>{% endif %}</li>
+            <li>{% if request.path == about_url %}Par skolotāju{% else %}<a href="{{ about_url }}">Par skolotāju</a>{% endif %}</li>
             {% url 'public:contacts' as contacts_url%}
-            <li>{% if request.path == contacts_url %}Kontakti{% else %}<a href={{ contacts_url }}>Kontakti</a>{% endif %}</li>
+            <li>{% if request.path == contacts_url %}Kontakti{% else %}<a href="{{ contacts_url }}">Kontakti</a>{% endif %}</li>
             {% url 'public:calendar' as calendar_url%}
-            <li>{% if request.path == calendar_url %}Kalendārs{% else %}<a href={{ calendar_url }}>Kalendārs</a>{% endif %}</li>
+            <li>{% if request.path == calendar_url %}Kalendārs{% else %}<a href="{{ calendar_url }}">Kalendārs</a>{% endif %}</li>
         </ul>
     </div>
     <div class="right">
         <ul>
             {% if user.is_authenticated %}
-                {% url 'accounts:profile' as profile_url %}
+                {% url 'public:profile' as profile_url %}
                 <li>Čau,
                     {% if request.path == profile_url %}{{user.username}}
                     {% else %}<a href="{{ profile_url }}">{{user.username}}</a>!{% endif %}</li>
-                <li><a href="{% url 'accounts:logout' %}">Iziet</a></li>
+                <li><a href="{% url 'public:logout' %}">Iziet</a></li>
             {% else %}
-                <li><a href="{% url 'accounts:login' %}">Ienākt</a></li>
+                <li><a href="{% url 'public:login' %}">Ienākt</a></li>
             {% endif %}
         </ul>
     </div>
