+++ sildapiano/sildapiano/urls.py
@@ -24,5 +24,5 @@ from . import views
 urlpatterns = [
     path("", include("sildapiano.apps.public.urls")),
     path("admin/", admin.site.urls),
-    path("accounts/", include("sildapiano.apps.accounts.urls")),
+    path("accounts/", include("sildapiano.apps.public.urls")),
 ]
